requirements.txt
tensorflow
streamlit 
matplotlib
numpy
scikit-image
huggingface_hub
pandas 
keras
Pillow
datasets
authlib
mysql-connector-python
